header/dialog break appart
