+++ client/my-sites/plans/components/domain-upsell-dialog/style.scss
@@ -1,31 +1,10 @@
-@import "@wordpress/base-styles/breakpoints";
-
-.is-section-plans .plans__section-header-skip {
-	display: flex;
-	justify-content: space-between;
-	margin: 0;
-	margin-bottom: 20px;
-
-	@media ( max-width: $break-small ) {
-		margin: 0 16px;
-	}
-}
-
-.is-section-plans .plans__section-header {
-	&.with-skip-button {
-		display: flex;
-		justify-content: space-between;
-		align-items: center;
-	}
-}
-
-.dialog.planspage__domain-upsell {
+.domain-upsell-dialog {
 
 	.dialog__content {
 		max-width: 572px;
 	}
 
-	.domain-upsell__modal-header {
+	.domain-upsell-dialog__modal-header {
 		text-align: right;
 
 		svg:hover {
