+++ client/my-sites/plans/components/plans-header/index.tsx
@@ -1,0 +1,92 @@
+import { isEnabled } from '@automattic/calypso-config';
+import { Button, Gridicon } from '@automattic/components';
+import { useCallback } from '@wordpress/element';
+import classNames from 'classnames';
+import { useTranslate } from 'i18n-calypso';
+import FormattedHeader from 'calypso/components/formatted-header';
+import { recordTracksEvent } from 'calypso/lib/analytics/tracks';
+import './style.scss';
+
+const DomainUpsellHeader: React.FunctionComponent< {
+	onOpenDialog: () => void;
+} > = ( { onOpenDialog } ) => {
+	const translate = useTranslate();
+	const is2023OnboardingPricingGrid = isEnabled( 'onboarding/2023-pricing-grid' );
+	const plansDescription = translate(
+		'See and compare the features available on each WordPress.com plan.'
+	);
+	const withSkipButton = ! is2023OnboardingPricingGrid;
+	const classes = classNames(
+		'plans__formatted-header',
+		'plans__section-header',
+		'modernized-header',
+		'header-text',
+		{
+			'with-skip-button': withSkipButton,
+		}
+	);
+
+	const onSkipClick = useCallback(
+		( event: React.MouseEvent< HTMLButtonElement, MouseEvent > ) => {
+			event.preventDefault();
+			recordTracksEvent( 'calypso_plans_page_domain_upsell_skip_click' );
+			onOpenDialog();
+		},
+		[ onOpenDialog ]
+	);
+
+	return (
+		<>
+			<FormattedHeader
+				className={ classes }
+				brandFont
+				headerText={ translate( 'Plans' ) }
+				subHeaderText={ plansDescription }
+				align="left"
+			>
+				{ withSkipButton && (
+					<Button className="plans__section-header-skip-button" onClick={ onSkipClick } href="/">
+						{ translate( 'Skip' ) }
+						<Gridicon icon="arrow-right" size={ 18 } />
+					</Button>
+				) }
+			</FormattedHeader>
+		</>
+	);
+};
+
+const PlansHeader: React.FunctionComponent< {
+	domainFromHomeUpsellFlow?: string;
+	onOpenDomainUpsellDialog: () => void;
+} > = ( { domainFromHomeUpsellFlow, onOpenDomainUpsellDialog } ) => {
+	const translate = useTranslate();
+	const plansDescription = translate(
+		'See and compare the features available on each WordPress.com plan.'
+	);
+
+	// const [ visibleDialog, setVisibleDialog ] = useState( false );
+
+	// const handleCloseDialog = useCallback( () => {
+	// 	setVisibleDialog( false );
+	// }, [ setVisibleDialog ] );
+
+	// const handleOpenDialog = useCallback( () => {
+	// 	setVisibleDialog( true );
+	// }, [ setVisibleDialog ] );
+
+	if ( ! domainFromHomeUpsellFlow ) {
+		return (
+			<FormattedHeader
+				className="plans__formatted-header plans__section-header modernized-header"
+				brandFont
+				headerText={ translate( 'Plans' ) }
+				subHeaderText={ plansDescription }
+				align="left"
+			/>
+		);
+	}
+
+	return <DomainUpsellHeader onOpenDialog={ onOpenDomainUpsellDialog } />;
+};
+
+export default PlansHeader;
