+++ client/my-sites/plans/main.jsx
@@ -43,8 +43,9 @@ import { getSelectedSite, getSelectedSiteId } from 'calypso/state/ui/selectors';
 import CalypsoShoppingCartProvider from '../checkout/calypso-shopping-cart-provider';
 import withCartKey from '../checkout/with-cart-key';
 import DomainAndPlanPackageNavigation from '../domains/components/domain-and-plan-package/navigation';
+import DomainUpsellDialog from './components/domain-upsell-dialog';
+import PlansHeader from './components/plans-header';
 import ECommerceTrialPlansPage from './ecommerce-trial';
-import PlansHeader from './header';
 import ModernizedLayout from './modernized-layout';
 
 import './style.scss';
@@ -205,6 +206,18 @@ class Plans extends Component {
 		return <ECommerceTrialPlansPage interval={ interval } siteSlug={ selectedSite.slug } />;
 	}
 
+	state = {
+		visibleDomainUpsellDialog: false,
+	};
+
+	handleCloseDomainUpsellDialog = () => {
+		this.setState( { visibleDomainUpsellDialog: false } );
+	};
+
+	handleOpenDomainUpsellDialog = () => {
+		this.setState( { visibleDomainUpsellDialog: true } );
+	};
+
 	render() {
 		const {
 			selectedSite,
@@ -217,6 +230,8 @@ class Plans extends Component {
 			isJetpackNotAtomic,
 			domainFromHomeUpsellFlow,
 		} = this.props;
+
+		const { visibleDomainUpsellDialog } = this.state;
 
 		if ( ! selectedSite || this.isInvalidPlanInterval() || ! currentPlan ) {
 			return this.renderPlaceholder();
@@ -254,10 +269,18 @@ class Plans extends Component {
 				<QueryContactDetailsCache />
 				<QueryPlans />
 				<TrackComponentView eventName="calypso_plans_view" />
+				<DomainUpsellDialog
+					visible={ visibleDomainUpsellDialog }
+					onClose={ this.handleCloseDomainUpsellDialog }
+					domain={ domainFromHomeUpsellFlow }
+				/>
 				{ canAccessPlans && (
 					<div>
 						{ ! isDomainAndPlanPackageFlow && (
-							<PlansHeader domainFromHomeUpsellFlow={ domainFromHomeUpsellFlow } />
+							<PlansHeader
+								domainFromHomeUpsellFlow={ domainFromHomeUpsellFlow }
+								onOpenDomainUpsellDialog={ this.handleOpenDomainUpsellDialog }
+							/>
 						) }
 						{ isDomainAndPlanPackageFlow && (
 							<>
